+++ Runtime/TypeScript/.eslintrc
@@ -1,0 +1,17 @@
+{
+  "root": true,
+  "parser": "@typescript-eslint/parser",
+  "plugins": ["@typescript-eslint", "prettier"],
+  "extends": [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended",
+    "prettier"
+  ],
+  "rules": {
+    "@typescript-eslint/no-non-null-assertion": 0,
+    "no-console": 1, // Means warning
+    "prettier/prettier": 2 // Means error
+  },
+  "ignorePatterns": ["src/generated/**/*.ts"]
+}
