+++ Runtime/TypeScript/.gitignore
@@ -1,2 +1,3 @@
 **.js
 dist/
+generated/
