+++ Runtime/TypeScript/tsconfig.json
@@ -3,12 +3,18 @@
         "target": "esnext",
         "module": "commonjs",
         "outDir": "dist",
+        "newLine": "lf",
+        "forceConsistentCasingInFileNames": true,
+        "strict": true,
+        "moduleResolution": "node",
+        "sourceRoot": "src",
+        "sourceMap": true,
         "declaration": true,
-        "moduleResolution": "node"
     },
     "exclude": [
         "node_modules",
         "dist",
         ".npm"
-    ]
+    ],
+    "include": ["src/**/*.ts"]
 }
